audio transfer to 2nd pi working, need to rethink pdm design.

+++ Pico_DSY_I2S_ADC_DMA/spi.h
@@ -11,7 +11,7 @@ extern "C"
 
 
 void init_spi();
-void spi_dma_transfer( float * this_buff, u32 buff_size_32_bit);
+void spi_dma_transfer( int16_t * this_buff, u32 buff_size_32_bit);
 
 extern float random_data[128];
